agents/main.py
import os
import json
import logging
import hashlib
from datetime import datetime
from typing import Optional, List, Dict
from fastapi import FastAPI, HTTPException, Body
from pydantic import BaseModel, Field
from dotenv import load_dotenv

load_dotenv()

# Logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger("lumifren-agents")

app = FastAPI(
    title="Lumifren Agents",
    description="Foundry-powered AI agents with semantic memory",
    version="1.0.0"
)

# ============================================================================
# MODELS
# ============================================================================

class CurateMemoryResponse(BaseModel):
    context: dict
    summary: str
    timestamp: str

class DetectToneResponse(BaseModel):
    tone: str
    confidence: float
    sentiment: str
    timestamp: str

class DevAssistantResponse(BaseModel):
    suggestions: list
    explanation: str
    priority: str
    timestamp: str

# ============================================================================
# AGENT LOGIC
# ============================================================================

@app.get("/health")
async def health():
    """Health check for the agent service"""
    return {"status": "ok"}

@app.post("/curate-memory", response_model=CurateMemoryResponse)
async def curate_memory(user_message: str = Body(...), user_id: str = Body(None)):
    """Curate semantic memory from user message."""
    logger.info(f"Curating memory for user {user_id}...")
    
    # Placeholder simulation of the Foundry logic
    context = {
        "extracted_entities": [],
        "key_topics": [],
        "emotional_state": "neutral",
        "memory_type": "semantic",
    }
    
    return {
        "context": context,
        "summary": f"Contextual summary for: {user_message[:100]}",
        "timestamp": datetime.utcnow().isoformat()
    }

@app.post("/detect-tone", response_model=DetectToneResponse)
async def detect_tone(user_message: str = Body(...), sentiment_analysis: str = Body("neutral")):
    """Detect tone and sentiment from user message."""
    logger.info(f"Detecting tone for: {user_message[:50]}...")
    
    # Simple simulation
    return {
        "tone": "neutral",
        "confidence": 0.8,
        "sentiment": "neutral",
        "timestamp": datetime.utcnow().isoformat()
    }

@app.post("/dev-assistant", response_model=DevAssistantResponse)
async def dev_assistant(code_context: str = Body(...), issue_description: str = Body(None)):
    """Provide development assistance."""
    logger.info(f"Dev assistance requested...")
    
    return {
        "suggestions": ["Refactor for clarity", "Add more tests"],
        "explanation": "The current code uses placeholders that should be replaced with real API calls.",
        "priority": "medium",
        "timestamp": datetime.utcnow().isoformat()
    }

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)
